refactor(maps): extract useMapDisplayedSpots from useMapFeaturesDraw
Move the selected/not-selected display-spot state and its two setters (setDisplayedSpots, setDisplayedSpotsWhileEditing) into a dedicated hook. No behavior change.
