+++ src/modules/maps/useMapFeaturesDraw.js
@@ -19,10 +19,11 @@ import {clearedVertexes, setFreehandFeatureCoords, setVertexStartCoords} from '.
 import useMapSymbology from './symbology/useMapSymbology';
 import useMap from './useMap';
 import useMapCoords from './useMapCoords';
+import useMapDisplayedSpots from './useMapDisplayedSpots';
 import useMapFeatures from './useMapFeatures';
 import useMapFeaturesCalculated from './useMapFeaturesCalculated';
 import useStereonet from './useStereonet';
-import {getNewId, getNewUUID, isEmpty, isEqual} from '../../shared/helpers';
+import {getNewId, getNewUUID, isEmpty} from '../../shared/helpers';
 import alert from '../../shared/ui/alert';
 import {setModalVisible} from '../home/home.slice';
 import {MODAL_KEYS} from '../page/pageKeys.constants';
@@ -63,7 +64,8 @@ const useMapFeaturesDraw = ({
 
   const {isDrawMode} = useMap();
   const {convertFeatureGeometryToImagePixels, convertImagePixelsToLatLong} = useMapCoords();
-  const {getAllMappedSpots, getDisplayedSpots} = useMapFeatures();
+  const {setDisplayedSpots, setDisplayedSpotsWhileEditing, spotsNotSelected, spotsSelected} = useMapDisplayedSpots();
+  const {getAllMappedSpots} = useMapFeatures();
   const {
     getDrawFeatureAtPress, getLassoedSpots, getSpotAtPress, getSpotsAtPress, identifyClosestVertexOnSpotPress,
   } = useMapFeaturesCalculated(mapRef);
@@ -80,8 +82,6 @@ const useMapFeaturesDraw = ({
   const [spotEditing, setSpotEditing] = useState({});
   const [spotsEdited, setSpotsEdited] = useState([]);
   const [spotsNotEdited, setSpotsNotEdited] = useState([]);
-  const [spotsNotSelected, setSpotsNotSelected] = useState([]);
-  const [spotsSelected, setSpotsSelected] = useState([]);
   const [vertexIndex, setVertexIndex] = useState([]);
   const [vertexToEdit, setVertexToEdit] = useState([]);
 
@@ -394,53 +394,6 @@ const useMapFeaturesDraw = ({
         'No Spots Selected',
       );
     }
-  };
-
-  // Set selected and not selected Spots to display when not editing
-  const setDisplayedSpots = (selectedSpots) => {
-    let [selectedDisplayedSpots, notSelectedDisplayedSpots] = getDisplayedSpots(selectedSpots);
-
-    // Convert image pixels to lat, lng (deep copy first to avoid mutating Redux state)
-    if (currentImageBasemap || stratSection) {
-      selectedDisplayedSpots = JSON.parse(JSON.stringify(selectedDisplayedSpots)).map(
-        spot => convertImagePixelsToLatLong(spot));
-      notSelectedDisplayedSpots = JSON.parse(JSON.stringify(notSelectedDisplayedSpots)).map(
-        spot => convertImagePixelsToLatLong(spot));
-    }
-
-    if (!isEqual(spotsSelected, selectedDisplayedSpots)) {
-      console.log('Selected Spots:', selectedDisplayedSpots);
-      setSpotsSelected(selectedDisplayedSpots);
-    }
-    if (!isEqual(spotsNotSelected, notSelectedDisplayedSpots)) {
-      console.log('Not Selected Spots:', notSelectedDisplayedSpots);
-      setSpotsNotSelected(notSelectedDisplayedSpots);
-    }
-  };
-
-  // Set selected and not selected Spots to display while editing
-  const setDisplayedSpotsWhileEditing = (spotEditingTmp, spotsEditedTmp, spotsNotEditedTmp) => {
-    if (!isEmpty(spotEditingTmp)) {
-      spotsNotEditedTmp = spotsNotEditedTmp.filter(spot => spot.properties.id !== spotEditingTmp.properties.id);
-    }
-    console.log('Set displayed Spots while editing. Editing:', spotEditingTmp, 'Edited:', spotsEditedTmp, 'Not edited:',
-      spotsNotEditedTmp);
-
-    let spotsEditedCopy = JSON.parse(JSON.stringify(isEmpty(spotsEditedTmp) ? [] : spotsEditedTmp));
-    let spotsNotEditedCopy = JSON.parse(JSON.stringify(isEmpty(spotsNotEditedTmp) ? [] : spotsNotEditedTmp));
-    let spotEditingCopy = JSON.parse(JSON.stringify(isEmpty(spotEditingTmp) ? [] : [{...spotEditingTmp}]));
-
-    // Convert image pixels to lat, lng
-    if (currentImageBasemap || stratSection) {
-      spotsEditedCopy = spotsEditedCopy.map(spot => convertImagePixelsToLatLong(spot));
-      spotsNotEditedCopy = spotsNotEditedCopy.map(spot => convertImagePixelsToLatLong(spot));
-      spotEditingCopy = spotEditingCopy.map(spot => convertImagePixelsToLatLong(spot));
-    }
-
-    console.log('Selected Edit Features:', spotEditingCopy);
-    setSpotsSelected(isEmpty(spotEditingCopy) ? [] : spotEditingCopy);
-    console.log('Unselected Edit Features:', [...spotsEditedCopy, ...spotsNotEditedCopy]);
-    setSpotsNotSelected([...spotsEditedCopy, ...spotsNotEditedCopy]);
   };
 
   const setEditFeatures = (spotToEdit) => {
